binary trees

+++ 9-binary_tree_height.c
@@ -9,7 +9,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
 {
 	if (tree)
 	{
-		size_t left = 0, right = 0;
+		int left = 0, right = 0;
 
 		if (tree->right)
 			right = 1 + binary_tree_height(tree->right);
